+++ kitty/cli.py
@@ -596,10 +596,8 @@ Set the name part of the :italic:`WM_CLASS` property (defaults to using the valu
 
 
 --title -T
-Set the OS window title. This will override any title set by the program running inside kitty. So
-only use this if you are running a program that does not set titles. If combined
-with :option:`{appname} --session` the title will be used for all windows created by the
-session, that do not set their own titles.
+Set the OS window title. This will override any title set by the program running inside kitty, permanently
+fixing the OS Window's title. So only use this if you are running a program that does not set titles.
 
 
 --config -c