+++ kitty/cocoa_window.m
@@ -574,12 +574,7 @@ cocoa_get_lang(PyObject UNUSED *self) {
         locale = [[NSLocale currentLocale] localeIdentifier];
     }
     if (!locale) { Py_RETURN_NONE; }
-    // Make sure the locale value is valid, that is it can be used
-    // to construct an actual locale
     const char* locale_utf8 = [locale UTF8String];
-    locale_t test_locale = newlocale(LC_ALL_MASK, locale_utf8, NULL);
-    if (!test_locale) { Py_RETURN_NONE; }
-    freelocale(test_locale);
     return Py_BuildValue("s", locale_utf8);
 
     } // autoreleasepool