+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 |kitty| is a feature-rich, cross-platform, *fast*, GPU based terminal.
 To update |kitty|, :doc:`follow the instructions <binary>`.
 
-0.23.0 [future]
+0.23.0 [2021-08-16]
 ----------------------
 
 - A new :doc:`themes kitten </kittens/themes>` to easily change kitty themes.